requirements.txt
numpy>=1.21.0
opencv-python>=4.5.3
scikit-learn>=0.24.2
scikit-image>=0.18.3
streamlit>=1.10.0
pillow>=8.3.2
matplotlib>=3.4.3
pandas>=1.3.3
tqdm>=4.62.3
joblib>=1.1.0 